requirements.txt
opencv-python==4.5.3.56
mediapipe==0.8.9.1
Pillow==8.4.0
numpy==1.21.4
scipy==1.7.3